Extract reusable components
