+++ Control/public/common/environment/informationRedirectActionPanel.js
